Bug: change implemenation to handle unsaved object
The original implementation to save the owner object on selecting
a record. However, for unsaved object this can lead to creating
the owner object twice! The implementation changed here to not
save the owner but update the has one relation, so that the main
form save button is required to save the relation.